Move SDK init span recording to SessionHandler (#204)
## Goal

Move the SDK init trace logging to SessionHandler because it is more reasonable to do it there. Technically doing it in `EmbraceImpl` is probably the most correct, but it's an OK spot here too and allows us to more explicitly test the relationship between `SessionHandler` and `EmbraceSpansService`

## Testing

Modified tests of `SpansServiceImpl` and reimplemented tests in the new `SessionHandler` test class.