Change release GH actions to create and release from release branch (#326)
## Goal

Create a release branch to make any version-specific modifications on and make the release from there. The version number of the SDK will incremented when the release branch is cut, and commits after that will be de facto in the next release. The Swazzler will be updated in the pre-release workflow as it does before.

If there are additional changes that need to go into the release branch, they can be done first in master, then cherrypicked into the release branch.

For hotfixes, the same cherrypicking process will be done, but we will need to add additional GH Actions to ensure to automate the release.