Updated pre-release and release workflow.
Improved steps, checkouts, etc.

We are now not pushing version to be released to remote master; because we don't care about that. Remote master only cares about having the next version to be released (-SNAPSHOT).

We are now publishing and closing Swazzler sonatype repository from this job.

The idea behind this is that we release the complete embrace suite from this job. Simple and easy.

Swazzler has been updated to work with nexus-plugin as well.