Fix flakey SpanTest by allow better validation on FakeSpanExporter (#401)
## Goal

Track the count of the numbers of spans exported and a method that waits for that value or times out. This fixes the flakiness for SpanTest but TracingApiTest is still flakey, which actually might be pointing to an actual race condition in the code that still needs to be addressed.

## Testing

Existing tests shouldn't be flakey