Change tracing API to take time in milliseconds (#449)
## Goal

Changing the public Embrace API to use milliseconds when specifying times to be more inline with Android developer expectations. Make every time-based parameter specify the unit in the name. Serialized forms stay in milliseconds to comply with OTel and not change server expectations.

The work to detect if a given timestamp is in nanoseconds and then auto converting to millis to ensure current usage don't fail will be done in a subsequent PR.

## Testing

Modified existing tests