Add detection and fallback to nanos of inappropriate times in millis (#454)
## Goal

We are changing the API to take in millisecond, but because the method signature hasn't changed, just how we are naming the parameter and how interpret it, there would be no compile time failure to warn folks they are passing in a time in nanos rather than millis.

For now, we'll work around this by detecting that a timestamp is way too big to be a reasonable time in millis in real life (until the year 5000), and assume its' a nanos time and convert it to millis. Once we detect that this isn't done anymore, we can remove this workaround.

## Testing

Added tests at all layers to verify this