Synchronise access when to files through the EmbraceCacheService (#471)
## Goal

Previously there was not any synchronisation when reading/writing files which in some cases could lead to incomplete, mangled, or empty payloads being persisted and sent. 

This approach is to add a `ReadWriteLock` in the `EmbraceCacheService`, as well as serialization writes (for the most part) on single thread to avoid excessive locking. A lock is allocated for each filename as this should always be unique & allows high throughput of other files. I chose `ReadWriteLock` as a further optimization - when a write is happening only one thread can hold a lock, but multiple threads are allowed to hold a read lock. As well, modifications were made so reads are done on the calling threads rather than on a background threads that then we do a blocking get on the resulting task.

Unit tests have been added to a component that emulates the locking semantics of `EmbraceCacheService`. Integration tests will be added later if you are reading this.