Write to a temp file when overwriting a session (#481)
## Goal

Change how we overwrite a session by first writing to a temp file, followed by renaming the old version, then the new one, then deleting the old version. This ensures that we will always have a valid copy of the session on disk no matter where the failure occurs.

In addition, handle cases where the session overwrite operation fails because the process terminates during one of the various steps, so when we load the cached sessions, we scan through the temp files to delete and rename as appropriate.

Also refactored the tests for `EmbraceDeliveryCacheManager` to not rely on a `TestCacheService` or check its internal operations when validating what is and isn't working.

## Testing

Added appropriate component tests and refactored the dependent tests.

To be added: tests to validate that serialization terminations will not leave partially written files. Moshi should guarantee that, but we shall see.