Update proguard config to fix R8 warnings introduced by using OTel logs SDK (#486)
## Goal

Running proguard on the test suite app caused some warnings that this fixed. Not sure if this is the best way to fix this, but the immutable annotation doesn't seem to be that important for this usage, but I'm open to better ideas. Simply keeping that doesn't seem to work