Add internal start method to enable the passing of a faked config service (#715)
## Goal

Much of the module overriding needed in the test harness is so a fake config service can be provided that is universally used. 

To simplify this, I've added a new internal method in `EmbraceImpl` that allows the passing of a ConfigService provider so that we can feed through one and get ModuleInitBootstrapper to initialize modules with a with it during integration tests.

It doesn't change the production code path for initialization, but I'm not super happy with having to do this. But ah well.