Separate URLConnection from OkHttp3 request interfaces
This is an initial set of changes with the objective of isolating deduplication logic only for URLConnection, as OkHttp3 does not need it.

The idea here was to clean the EmbraceImpl interfaces, give them clearer names, and move logic that will decide if deduplication is needed to NetworkLoggingService, where we will use DataSources in later commits.