Initial prep for sending network requests as spans (#745)
## Goal

This is the initial PR for sending network requests as spans. 

- Removed recordAndDeduplicateNetworkRequest. 

This was used in HttpURLConnectionDelegate to avoid logging network requests with the same call ID. It is still required, but we will address it differently in the following PRs. For now, we will use the same method as the rest of the requests.

- logNetworkError and logNetworkCall did essentially the same. We will move that logic to the logging service so we don't bloat EmbraceImpl

- getNetworkCallsSnapshot won't be used anymore, as we will record spans directly when they get to the logging service. We won't need to expose them when the session ends.

- Stop looking for network logs in PerformanceInfo. Removed "nr" from session-end.json. We won't be sending network requests there anymore.

- Updated tests to use the new APIs