Record logged network requests as completed spans (#748)
- Record network requests as spans.

- Consolidate logNetworkCall and logNetworkError methods

EmbraceNetworkLoggingService had two methods: logNetworkCall and logNetworkError. They were both doing essentially the same, but with a different NetworkCallV2 object: strip the URL, sanitize the TraceId, logging the networkCapturedData, and processing the network call.

We don't need a NetworkCallV2 object anymore, and the EmbraceNetworkRequest is the same for both network call and network error, so we can just use one method.

- Remove unused methods.

### Questions:
- Why do we need to strip the URL? 
- Why are we stripping the URL for the span, but not for the networkCapturedData?