Fixing logRnView (#924)
* Remove unused force flag from changeView in ViewDataSource

* Remove unused timestamp parameter from logView in EmbraceBreadcrumbService

* Removed isAutomaticActivityCaptureEnabled config check from the ViewDataSource creation

We only need to check for this when we call it from the ActivityLifecycleTracker. It was conflicting with the logRnView method.

* Fixed LogRNView and added tests

If you call logRnView with the same view name, a new span will be created with the same view name.

* Brought back timestamps for logView as we will use them later

(cherry picked from commit bca00d2aa69ac4e9cf41865394e1bb611037f0ff)