Create session span when one is expected and not present (#1166)
## Goal

A recent change made it so that session spans aren't automatically created when a session stops. However, because background activity can be disabled, we could be in a state where there is no active session. In those cases, when we foreground, we have to create that session span.

This change ensures that when a payload is constructed for a new session, a session span is created if it doesn't already exist.

## Testing

Tests are added in all surface areas for which this is relevant. Additional integration tests have been added as well to ensure session transitions result in the right state depending on whether background activities are enabled.