Always start a new session span after a manual session end (#1167)
## Goal

Manual session ends only started a new session span if background activity is enabled. This is because there's ambiguity in how we decided whether we should new session span.

To address this, I've consolidated the decision in the `FinalEvelopeParam` definition and removed it from the `SessionSnapshotType`, what is defined in FinalEvelopeParam is the ultimate arbitor.

## Testing

Added additional testing to verify the correctness at the PayloadFactory level as well as the PayloadMessageCollator level.