Increase span attribute limit at the OTel layer (#1203)
## Goal

OTel imposes a limit on how many attributes are allowed on a span, which is 128 by default. For the session span, we require more than that because you can theoretically have 100 session properties.

To support this, we bump the total limit set in the OTel layer to 300, and validate that our custom limit of 50 for SDK users is not enforced when the SDK adds attributes via the internal `addSystemAttribute()` method. This will also bump the total number of other attributes the SDK can add to the session span to a safer ceiling.

## Testing

Added and modified appropriate unit tests to ensure the expected maximums can be written.