Populate session properties in session span during transition (#1223)
## Goal

We were populating session properties in a session span as a side effect of constructing the end payload of the previous session. This caused a couple of problems: session-only properties not being cleared before the population, and if background activity is disabled, session properties population fails.

Changing where we do the population fixes both of these, as well as allow us to remove the extra population call during SDK startup. 

This change also supersedes #1206 as literal removal is no longer necessary - session only properties will be added or removed from the current session's session span by the `SessionPropertiesService` 

## Testing

Add integration tests for all cases, as well as the one added in #1206 to verify that session-only props don't bleed into the next session. Unit tests were also modified accordingly.