Expose deviceId as a synthetic immutible property (#1258)
## Goal

Previous SDK versions had getDeviceId() exposed as a synthetic property on the Java `Embrace` object in Kotlin, but it's no longer true, likely because of the Kotlin version change. Restored this ability by exposing it as an actual properly on the interface.

## Testing
Added additional testing for the case when the SDK is not started. This doesn't change how part behaved, but codecov yelled at me, so I obliged.