Allow getLastRunEndState() to be accessed via a synthetic property in Kotlin (#1282)
## Goal

Store the ability to access `getLastRunEndState()` via a synthetic read-only property in Kotlin as it was that way in previous versions thanks to the old Kotlin version that was used. 

We will revisit this in the next major version.

## Testing
Add integration tests that access the getter in the Java Embrace object via this