Allow internal spans to have longer names (#1354)
## Goal

Allow spans that the SDK logs as its out of the box instrumentation to have longer names as well as attribute values. This solves the problem of network requests having long names and attributes because of the URL being logged is greater than what was set as the ceiling for perf traces.

## Testing
Add unit tests and integration tests