Simplify session properties tests to not checked cached payload (#1412)
## Goal

Getting the last saved session was flakey in the same way getting the last saved background activity was flakey. Fixed this by just not checking that, but instead checking when the session is done, which is not flakey because that save happens on the main test thread.

At the same time, I made the non-flakey saved envelope payload validation function generic and usable for sessions, if we ever need it