Fake app lifecycle when faking a session during integration tests (#2019)
## Goal

Make it so that session creation in integration tests will simulate the app/activity lifecycle of the app. This ensures that side effects from lifecycle handlers are taken into account when running the tests end to end.

To that end, I've adding more functionality to the `recordSession()` method, having it call `simulateOpeningActivities()` with the appropriate parameters, and returning an object that contains some useful timestamps to validate against.

This requires a few changes to the tests:

- Change time verifications now that the clock isn't only moved manually 
- Account for the extra telemetry that are logged as a result of going through the lifecycle
- Tweak some internals of the faking logic to ensure that the callbacks are invoked in the same order as in prod