Adds support for render detect for Android 6+ (#2027)
## Goal

Use an alternate way to detect that a frame was been delivered that works for Android 12 and Android 6 to 9. Implement a new `DrawEventEmitter` that uses this alternative method, and supply when the other one doesn't work.

All the unit tests have been updated to reflect this new reality, but the integration tests don't check for hardware based render events. But with this, we may be able to test this by simulating this, so stay tuned!

## Testing
Manually tested for a subset of real devices/apps. Will test it more comprehensively.