Rename less good version of render span for clarity (#2033)
## Goal

Rename app startup render child span that tracks less. An alternative implementation is adding a new event to be tracked, and keying off that to create the span.

I don't like that because a new event adds little value for android versions that can do frame commit callbacks, and it adds complication to the code. Basically using one or the other and calling both of them render removes that distinction, so we just have to make sure the name is correct. I prefer this.