Ensure ANR spans are exported when a session ends (#2136)
## Goal

Log ANRs as real spans through the SDK if we are producing a final session payload. This allows for ANRs to be exported as spans when a session ends. To that end, I added a new method to take the `AnrInterval` objects and log them as real spans that will be used when a session is ending, and only use the method to produce the fake spans to stuff in a session payload snapshot before it ends when we are doing a snapshot. 

## Testing

Existing tests validate that the logic change doesn't affect the final payload. Additional tests added to validate the exported spans.