Increase limit for log properties to 100 (#2392)
## Goal

Update limits for attribute key length, value length, and counts to 128, 1024, and 100, respectively. Add a cap of 200 for remote configured session prop limits as well.

`PropertyUtil` was also refactored so it contains a truncation method, but the sanitization method was inlined into the `EmbraceLogService` class because that's the only place that used it. Tests were ported too - at least the relevant ones.