Move span event validation so truncation (#2404)
## Goal

Events were dropped if they had names that were too long or if they contained too many attributes. Attributes were dropped if their keys or values were too long. This adds similar truncation logic, and moves the check out of the `EmbraceSpanEvent` factory method, which was a terrible idea to begin with. Now, it basically just creates the object and allows the truncation to happen at a lower layer.

In addition, the span name used by the created span is now wholly inside `EmbraceSpanImpl`. This is so that the arguments object's name can be immutable so truncation can be localized inside the `EmbraceSdkSpan` implementation.