Suppress the logcat warning when applicationInitEnd called before SDK starts (#2442)
## Goal

Because we bytecode instrument the `applicationInitEnd()` into every `Application` class, if an app explicitly calls the `onCreate()` of a a super class, `applicationInitEnd()` could be invoked when the app's Application class' `onCreate()` is called, i.e. before the Embrace SDK starts. Therefore, we'll suppress this misleading logcat message when `applicationInitEnd()` is called before the SDK starts.