Enable Embrace SDK to use the Kotlin SDK (#2487)
## Goal

Add in a constant that defines which SDK implementation the Embrace SDK will use. Right now, there's no way to flip this at runtime, so it currently behaves as it would before. The main changes in this PR is putting in the plumbing and ensuring all tests pass when the switch is flipped.

To do this, we require changing assumptions in the SDK and tests of the compat implementation always being used, as well as well as make a handful of components, test helpers, and fakes be aware of the difference and act accordingly.

In practice, there are only a small number of places in the production code that need to be aware of this, and it's all in the otel module in terms of which SDK to initialize and how to get to default parent of a span. Kudos to the team to design and implement the architecture in such a way that this is possible!

Right now,, a few tests have been changed so it always tests using the compat implementation, but most will figure out which implementation is used by looking at a parameter or the constant flag, the latter minimized as much as possible.