Consolidate code solely needed by HUC instrumentation (#2601)
## Goal

Clean up HUC to make extraction into a separate module easier. This involved converting `HttpUrlConnectionTracker`​ into Kotlin and consolidated related logic into to, much of which were exposed via difficult-to-test Java static objects and methods, because that's what you have to do to work with the HUC API if you want to do what we are doing. Other static helpers were converted into Kotlin and moved out of the `embrace-android-sdk` module as well, while leaving much of the HUC instrumentation unchanged (and in Java).