Add instrumentation to HUC wrapper (#2635)
## Goal

Create a more limited version of HUC instrumentation that doesn't support the advanced features of the old one (`embrace-android-instrumentation-huc`​), but is less invasive in how it does the instrumentation because it doesn't need to wrap streams, count bytes, and read and inject  headers. Therefore, it's more appropriate to be used to track requests to third-party APIs that use this network stack as a fallback, so it will be the new default if no configurations are set. As it doesn't support HTTP/2, it's not recommended to be used for first-party APIs because of the inherent perf penalties, though if someone were to use it, they can use the old instrumentation to track everything as before.

This new lite HUC instrumentation will be enabled by default.