Refactor HUC Lite instrumentation to use TelemetryDestination (#2653)
## Goal

Convert HUC Lite instrumentation to a `DataSource`​ and initialize via SPI. Some clean up and optimization still need to be done, but this should be functional, so I want to get it reviewed first.  
  
Note that we still need to enforce limits - this has to be shared with all the other network requests, so it will likely have to be plumbed through the args that come through during registration. But I thought I'd get this in review