Do session transition inside of SessionTracker (#2911)
## Goal

The first of step of rationalizing session bookkeeping and create a better defined session lifecycle where instrumentation can hook into its various parts is to consolidate the logic of starting, ending, and transitioning between sessions in one place: `SessionTracker`

This PR restructures the existing transitioning by largely keeping the code path the same. `SessionOrchestrator` and `SessionTracker` has also been refactored so that the latter contains the source of truth when it comes to what the active session is.

A future PR will add a hook so instrumentation can "do something" before a session ends so that state spans can be finalized within them. More refactoring can be done after that, but lets get that in first.

<!-- Describe what this change seeks to address -->

## Testing

Existing tests should cover this